Update README.md
There can still be regulations and growing limitations but that should not hinder the overall revenue the cannabis industry will experience.  Canada welcomed cannabis with open-mindedness and the best is yet to come with the onset of greater possibilities for expansion.  If you are on the look-out for the top of the line cannabis products at an affordable price, Sativa Bliss is your both physical and virtual go-to store.  Contact them easily through their social media accounts or better check their website at www.sativabliss.ca.